fix: fixed all comments
